job_control: stop leaking our file descriptors into spawned children (leader, workers, tunnels)
Core pid 1096123 (v1.2.16, 2026-08-23): SIGSEGV in wxEpollDispatcher::Dispatch
calling through an already-freed wxSocketImpl. wx sockets are not close-on-exec,
so the non-compute leader (and every run command / tunnel) spawned with system()
inherited a dup of job_control's LISTENING socket. When the shutdown gate closed
the listener, the kernel kept the socket and our epoll registration alive via the
leader's dup; the worker shim's 10 s watchdog probe then completed a handshake
against that zombie listener, EPOLLIN fired on the stale entry, and the event loop
dereferenced freed memory (crash 10 s after the gate, every leader-on run).

MarkAllFdsCloseOnExec() sets FD_CLOEXEC on fds >= 3 before each spawn; all
system()/popen() sites now go through SystemNoInherit/PopenNoInherit. The OnInit
log mirror keeps plain system() on purpose (needs the dup'd stdout; no sockets
exist yet). Also cast the last long-to-%i in the myapp worker-connect trace.

+++ src/core/myapp.cpp
@@ -1290,7 +1290,7 @@ void MyApp::HandleNewSocketConnection(wxSocketBase* new_connection, unsigned cha
         // expected connection is a real worker.
         int expected_worker_connections = i_am_a_non_compute_leader ? number_of_commands_to_run : number_of_commands_to_run - 1;
         wxPrintf("MASTER: worker socket %p connected (%i of %i expected)\n", (void*)new_connection, (int)worker_socket_pointers.GetCount( ), expected_worker_connections);
-        MyDebugPrintWithDetails("LEADER-PATH: worker connected %p (%i of %i), max_connected=%i dispatched=%i\n", (void*)new_connection, (int)worker_socket_pointers.GetCount( ), expected_worker_connections, max_number_of_connected_workers, number_of_dispatched_jobs);
+        MyDebugPrintWithDetails("LEADER-PATH: worker connected %p (%i of %i), max_connected=%i dispatched=%i\n", (void*)new_connection, (int)worker_socket_pointers.GetCount( ), expected_worker_connections, (int)max_number_of_connected_workers, (int)number_of_dispatched_jobs);
         if ( worker_socket_pointers.GetCount( ) == expected_worker_connections ) {
             SocketSendInfo("All workers have re-connected to the master.");
         }
